Make hooks able to access all fields
Previously, because we set the serializer fields based on the permissions, it meant the hooks could not set fields on the model that were not in the statezero configured fields. This was problematic, so we have fixed it. Now the user data is filtered based on the permissions, but the serializer for writes we construct dynamically contains all the fields.